Better example in the readme
This is a cool action :) I struggled a bit with errors, but have figured them out and thought I should open a PR to update your README.

The current readme doesn't actually work, as it doesn't include the reference to the `GITHUB_TOKEN` env. I also think that for most people, this is going to need to be executed when the title of a PR is updated, so I updated the example triggers as well.